Revert to the code from 0.0.2: Works
